requirements.txt
asgiref==3.7.2
Django==4.2.9
django-dump-load-utf8==0.0.4
django-extensions==3.2.3
djangorestframework==3.14.0
graphviz==0.20.1
pillow==10.2.0
pydot==2.0.0
pyparsing==3.1.1
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
